% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
A data frame with 2 variables:
\describe{
  \item{DISTANCE}{Distance from the building to the nearest fire station (in miles)}
  \item{DAMAGE}{Estimated fire damage (in dollars)}
}
}
\source{
MATH 4753 Lab 7 assignment dataset.
}
\usage{
fire
}
\description{
A dataset containing distance to the nearest fire station and the
amount of fire damage recorded for a set of buildings.
}
\keyword{datasets}
