% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations.}

\item{x}{Numeric vector of data.}

\item{fun}{Statistic to compute (e.g., "mean", "var", median, IQR).}

\item{alpha}{Level of significance.}

\item{cx}{Text scaling for labels.}

\item{...}{Additional graphical parameters for hist().}
}
\value{
A list containing:
\item{ci}{Bootstrap confidence interval}
\item{xstat}{Bootstrap statistics}
\item{fun}{Statistic used}
\item{x}{Original sample}
}
\description{
This function performs bootstrap resampling and constructs
confidence intervals for a user-specified statistic.
}
\examples{
# myboot2(x = rnorm(20))

}
