% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Airline Overbooking Problem Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Integer. Number of seats on the plane.}

\item{gamma}{Numeric. Target overbooking probability (e.g., 0.02 for 2 Percent).}

\item{p}{Numeric. Probability a passenger shows up (e.g., 0.95).}
}
\value{
A named list with the following components:
  \describe{
    \item{nd}{Number of tickets using the discrete Binomial method}
    \item{nc}{Number using the Normal approximation}
    \item{N}{Input number of seats}
    \item{p}{Probability of showing up}
    \item{gamma}{Overbooking probability}
  }
}
\description{
Calculates the number of tickets an airline should sell when the number of
seats is N, the probability of a show-up is p, and gamma is the maximum
acceptable probability of overbooking (too many people showing up).
}
\details{
The function also creates plots showing the objective function for both
discrete and normal approximation cases.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)

}
